backend/requirements.txt
attrs==22.1.0
beautifulsoup4==4.11.2
cbor==1.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Cython==0.29.33
exceptiongroup==1.0.4
ez-setup==0.9
fasttext==0.9.2
filelock==3.9.0
Flask==2.2.2
fst-pso==1.8.1
FuzzyTM==2.0.5
gensim==4.3.0
glove==1.0.0
huggingface-hub==0.12.1
idna==3.4
ijson==3.2.0.post0
importlib-metadata==5.1.0
iniconfig==1.1.1
ir-datasets==0.5.4
itsdangerous==2.1.2
Janome==0.4.2
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.2
lz4==4.3.2
MarkupSafe==2.1.1
maven==0.0.6
mecab-python3==1.0.7.dev1+g6a1ecfd
miniful==0.0.6
nltk==3.8.1
numpy==1.23.5
packaging==21.3
pandas==1.5.2
pluggy==1.0.0
psycopg2-binary==2.9.5
pyautocorpus==0.1.9
pybind11==2.10.3
pyFUME==0.2.25
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.97
simpful==2.10.0
sister==0.1.10
six==1.16.0
smart-open==6.3.0
soupsieve==2.4
threadpoolctl==3.1.0
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
tqdm==4.64.1
transformers==4.26.1
trec-car-tools==2.6
typing_extensions==4.5.0
unidic-lite==1.0.8
unique-names-generator==1.0.2
unlzw3==0.2.2
urllib3==1.26.13
warc3-wet==0.2.3
warc3-wet-clueweb09==0.2.5
Werkzeug==2.2.2
xlrd==2.0.1
youtube-transcript-api==0.5.0
zipp==3.11.0
zlib-state==0.1.5
